--- client/src/components/Blog/Courses/CourseList/Content/Modals/AddUser.tsx
@@ -1,38 +1,0 @@
-import {
-    Button,
-    Modal,
-    ModalContent,
-    ModalHeader,
-    ModalBody,
-    ModalFooter,
-  } from "@nextui-org/react";
-  import { AddUserForm } from "./Forms/AddUserForm";
-  
-  type AddUserModalType = {
-    isOpen: boolean;
-    onOpenChange: (open: boolean) => void;
-  };
-  
-  export const AddUser = ({ addUser }: { addUser: AddUserModalType }) => {
-    return (
-      <Modal
-        isOpen={addUser.isOpen}
-        onOpenChange={addUser.onOpenChange}
-        placement="center"
-        backdrop="blur"
-        scrollBehavior="inside"
-        radius="sm"
-        className="!bg-[#f5f1ec] border-1 shadow-lg"
-      >
-        <ModalContent>
-          <ModalHeader>
-            Dodaj Użytkownika
-          </ModalHeader>
-          <ModalBody className="mt-0 pt-0">
-            <AddUserForm onClose={() => addUser.onOpenChange(false)} />
-          </ModalBody>
-        </ModalContent>
-      </Modal>
-    );
-  };
-
