+++ client/src/components/Blog/Courses/CourseList/Content/module.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import { EditCourseForm } from "./Modals/EditCourse";
-import AddCourseForm from './Modals/AddCourse';
+import React from "react";
+import EditCourseModal from "./Modals/EditCourse";
+import AddCourseForm from "./Modals/AddCourse";
 import {
   Button,
   Modal,
@@ -14,8 +14,11 @@ import {
 import { Card, CardHeader, CardBody } from "@nextui-org/react";
 import { Link } from "@nextui-org/react";
 import { useSession } from "next-auth/react";
-import { useCoursesQuery, useDeleteCourseMutation } from "@/services/courseHooks"; 
-import { Course, CoursesResponse } from '@/components/Home/ts/types'; 
+import {
+  useCoursesQuery,
+  useDeleteCourseMutation,
+} from "@/services/courseHooks";
+import { Course, CoursesResponse } from "@/components/Home/ts/types";
 
 export default function Test1() {
   const { data: session, status } = useSession(); // Use useSession to manage session state
@@ -23,13 +26,9 @@ export default function Test1() {
   const editModal = useDisclosure();
   const detailModal = useDisclosure();
   const [selectedCourse, setSelectedCourse] = React.useState<Course>();
-  const isAdmin = session?.user?.role === 'administrator'; // Sprawdzenie, czy użytkownik jest administratorem
+  const isAdmin = session?.user?.role === "administrator"; // Sprawdzenie, czy użytkownik jest administratorem
 
-  const {
-    data: CourseResponse,
-    error,
-    isLoading,
-  } = useCoursesQuery(); // Fetch courses
+  const { data: CourseResponse, error, isLoading } = useCoursesQuery(); // Fetch courses
   const deleteCourseMutation = useDeleteCourseMutation({
     onSuccess: () => {
       // Odśwież listę kursów lub wykonaj inne akcje po sukcesie
@@ -40,7 +39,7 @@ export default function Test1() {
     },
   });
   const courses = CourseResponse?.data || [];
-  
+
   const handleDeleteCourse = (courseId: string) => {
     console.log(courseId); // Dodaj to, aby upewnić się, że ID jest poprawne
     if (confirm("Czy na pewno chcesz usunąć ten kurs?")) {
@@ -49,13 +48,24 @@ export default function Test1() {
     }
   };
 
-  const handleCourseClick = (course:Course) => {
+  const handleCourseClick = (course: Course) => {
     setSelectedCourse(course);
     detailModal.onOpen();
   };
 
   const CourseIcon = () => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="lucide lucide-book-plus">
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className="lucide lucide-book-plus"
+    >
       <path d="M12 7v6" />
       <path d="M4 19.5v-15A2.5 2.5 0 0 1 6.5 2H19a1 1 0 0 1 1 1v18a1 1 0 0 1-1 1H6.5a1 1 0 0 1 0-5H20" />
       <path d="M9 10h6" />
@@ -73,62 +83,76 @@ export default function Test1() {
             </span>
           </h3>
           <p className="text-sm text-pretty leading-relaxed text-gray-700 text-justify text-clip">
-            Oto lista dostępnych szkoleń. {isAdmin ? 
-              "Jako administrator, masz pełną kontrolę nad tymi zasobami – możesz dodawać nowe szkolenia, edytować istniejące oraz usuwać te, które są już nieaktualne lub niepotrzebne." :
-              "Jako użytkownik możesz przeglądać dostępne szkolenia i zarejestrować się na te, które Cię interesują."
-            }
+            Oto lista dostępnych szkoleń.{" "}
+            {isAdmin
+              ? "Jako administrator, masz pełną kontrolę nad tymi zasobami – możesz dodawać nowe szkolenia, edytować istniejące oraz usuwać te, które są już nieaktualne lub niepotrzebne."
+              : "Jako użytkownik możesz przeglądać dostępne szkolenia i zarejestrować się na te, które Cię interesują."}
           </p>
         </div>
         <Divider className="h-[1px] w-full" />
-      </div>
+        <div className="flex flex-col gap-3">
+          {isAdmin && (
+            <div className="flex flex-col sm:flex-row gap-3 justify-end items-center">
+              <Button
+                onPress={addModal.onOpen}
+                className="w-full lg:w-auto rounded text-white bg-green-700"
+                endContent={<AddIcon />}
+              >
+                Dodaj Szkolenie
+              </Button>
+              <AddCourseForm addModal={addModal} />
+            </div>
+          )}
 
-      <div className="flex flex-col gap-3">
-        {isAdmin && (
-          <div className="flex flex-col sm:flex-row gap-3 justify-end items-center">
-            <Button
-              onPress={addModal.onOpen} 
-              className="w-full lg:w-auto rounded text-white bg-green-700"
-              endContent={<AddIcon />}>
-              Dodaj Szkolenie
-            </Button>
-            <AddCourseForm addModal={addModal} />
-          </div>
-        )}
-
-        {courses.length > 0 ? (
-          <div className="gap-6 grid grid-cols-1 lg:grid-cols-2 xl:grid-cols-4">
-            {courses.map(course => (
-              <Card key={course.id} shadow="lg" className="bg-[#f5f1ec] border-2 border-[#333]/25 rounded">
-                <CardHeader className="flex flex-col justify-start sm:flex-row items-start gap-3">
-                  <div className="flex flex-col gap-2">
-                    <CourseIcon />
-                    <div className="flex flex-col">
-                      <p className="text-small font-medium text-justify">{course.name}</p>
-                      <p className="text-small text-default-500">Szkolenie #{course.id}</p>
+          {courses.length > 0 ? (
+            <div className="gap-6 grid grid-cols-1 xl:grid-cols-2">
+              {courses.map((course) => (
+                <Card
+                  key={course.id}
+                  shadow="lg"
+                  className="bg-[#f8f7f5] border-2 border-[#333]/25 rounded"
+                >
+                  <CardHeader className="flex flex-col justify-start sm:flex-row items-start gap-3">
+                    <div className="flex flex-col gap-2">
+                      <CourseIcon />
+                      <div className="flex flex-col">
+                        <p className="text-small font-medium text-justify">
+                          {course.name}
+                        </p>
+                        <p className="text-small text-default-500">
+                          Szkolenie #{course.id}
+                        </p>
+                      </div>
                     </div>
-                  </div>
-                </CardHeader>
-                <Divider className="h-[1px] w-full" />
-                <CardBody>
-                  <Button className="w-full lg:w-auto rounded text-white bg-green-700 text-small text-justify" onClick={() => handleCourseClick(course)} >Sprawdź Informacje</Button>
-                </CardBody>
-              </Card>
-            ))}
-          </div>
-        ) : (
-          <></>
-        )}
+                  </CardHeader>
+                  <Divider className="h-[1px] w-full" />
+                  <CardBody>
+                    <Button
+                      className="w-full lg:w-auto rounded text-white bg-green-700 text-small text-justify"
+                      onClick={() => handleCourseClick(course)}
+                    >
+                      Sprawdź Informacje
+                    </Button>
+                  </CardBody>
+                </Card>
+              ))}
+            </div>
+          ) : (
+            <></>
+          )}
+        </div>
       </div>
 
-      <Modal 
+      <Modal
         placement="center"
         backdrop="blur"
         scrollBehavior="inside"
         radius="sm"
-        size='5xl'
-        className="!bg-[#f5f1ec] border-1 shadow-lg " 
-        isOpen={detailModal.isOpen} 
-        onClose={detailModal.onClose}>
+        size="5xl"
+        className="!bg-[#f5f1ec] border-1 shadow-lg "
+        isOpen={detailModal.isOpen}
+        onClose={detailModal.onClose}
+      >
         <ModalContent>
           <ModalHeader>
             {selectedCourse ? selectedCourse.name : "Szczegóły Kursu"}
@@ -136,12 +160,26 @@ export default function Test1() {
           <ModalBody>
             {selectedCourse && (
               <div className="flex flex-col gap-3 text-pretty leading-relaxed">
-                <p className='flex flex-col'><strong>Opis Szkolenia:</strong> {selectedCourse.description}</p>
-                <p className='flex flex-col'><strong>Data Szkolenia:</strong> {selectedCourse.date}</p>
-                <p className='flex flex-col'><strong>Zakończenie Szkolenia:</strong> {selectedCourse.endDate}</p>
-                <p className='flex flex-col'>
+                <p className="flex flex-col">
+                  <strong>Opis Szkolenia:</strong> {selectedCourse.description}
+                </p>
+                <p className="flex flex-col">
+                  <strong>Data Szkolenia:</strong> {selectedCourse.date}
+                </p>
+                <p className="flex flex-col">
+                  <strong>Zakończenie Szkolenia:</strong> {selectedCourse.endDate}
+                </p>
+                <p className="flex flex-col">
                   <strong>Plik Szkoleniowy:</strong>
-                  <Link href={selectedCourse.course_link} color="success" className='text-green-800' size='sm'> Plik do Pobrania</Link>
+                  <Link
+                    href={selectedCourse.course_link}
+                    color="success"
+                    className="text-green-800"
+                    size="sm"
+                  >
+                    {" "}
+                    Plik do Pobrania
+                  </Link>
                 </p>
               </div>
             )}
@@ -152,17 +190,26 @@ export default function Test1() {
                 <Button
                   color="success"
                   variant="flat"
-                  className="rounded w-full"
-                  onPress={editModal.onOpen}>
+                  className="rounded w-full bg-green-700 !text-white"
+                  onPress={editModal.onOpen}
+                >
                   Edytuj
                 </Button>
                 {/* Fixed courseId */}
-                <EditCourseForm courseId={selectedCourse?.id || 0} editModal={editModal} />
+                <EditCourseModal
+                  courseId={selectedCourse?.id || 0}
+                  isOpen={editModal.isOpen} 
+                  onOpenChange={editModal.onOpenChange} 
+                  onOpen={() => editModal.onOpen()} 
+                  onClose={() => editModal.onClose()} 
+                />
                 <Button
                   color="danger"
                   variant="flat"
                   className="rounded w-full"
-                  onClick={() => handleDeleteCourse(selectedCourse?.id.toString() || "0")}
+                  onClick={() =>
+                    handleDeleteCourse(selectedCourse?.id.toString() || "0")
+                  }
                 >
                   Usuń
                 </Button>
@@ -176,7 +223,18 @@ export default function Test1() {
 }
 
 const AddIcon = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="lucide lucide-book-plus">
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    className="lucide lucide-book-plus"
+  >
     <path d="M12 7v6" />
     <path d="M4 19.5v-15A2.5 2.5 0 0 1 6.5 2H19a1 1 0 0 1 1 1v18a1 1 0 0 1-1 1H6.5a1 1 0 0 1 0-5H20" />
     <path d="M9 10h6" />
